move buildSearchRequest back into actions file
https://coveord.atlassian.net/browse/KIT-317
